Moved create a conference code to a separate file
